manejo-documentos/worker/requirements.txt
google-cloud-pubsub==2.*
google-cloud-storage==2.*
